% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_similarity.R
\name{grouping_similarity}
\alias{grouping_similarity}
\title{Comparing similarity of two (or more) groupings}
\usage{
grouping_similarity(gr1, gr2 = NULL, groupwise.similarity = FALSE)
}
\arguments{
\item{gr1}{a vector or a data frame (each columns should indicate different groupings)}

\item{gr2}{a vector to compare \code{gr1}, Default: NULL}

\item{groupwise.similarity}{Should we also show the similarity by each group?, Default: FALSE}
}
\value{
\{wholelevel.similarity} A matrix showing the similarity of two groupings (ranging from 0 to 1).<br>
\{groupwise.similarity} Similarity of each group in gr1 (if the input is the vector) or the first column
(if the input is a data frame) and the rest of the groupings.
}
\description{
\code{grouping_similarity} calculates the similarity of two (or more) groupings (e.g. respondents).
This function can be used, for example, to check how similar different clustering solutions are.
}
\details{
Based on the groupings in \code{gr1} and \code{gr2},
\code{grouping_similarity} first construct the matrices in that the entry
indicates 1 if i and j belong to the same group and 0 otherwise.
It then calculate the standardized hamming distance among matrices
(i.e., the realized hamming distance divided by the maximum hamming distance).
}
