% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usagi-package.R
\docType{package}
\name{usagi-package}
\alias{usagi}
\alias{usagi-package}
\title{usagi: Utilities for Sociological Analysis with Getting (good) Ideas}
\description{
A set of utility functions for conducting a Sociological analysis, especially for exploratory analysis. This includes tools for descriptive statistics, clustering, and social network analysis.
}
\author{
\strong{Maintainer}: Keiichi Satoh \email{keiichi.satoh@r.hit-u.ac.jp} (\href{https://orcid.org/0000-0003-3316-301X}{ORCID})

}
\keyword{internal}
